# client/src/api/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ShoppingServer",
    "version": "1.0"
  },
  "paths": {
    "/api/AddAddress": {
      "post": {
        "tags": ["Address"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAddressOperationInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAddressOperationInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddAddressOperationInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddAddressOperationOutputDtoOperationOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddAddressOperationOutputDtoOperationOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddAddressOperationOutputDtoOperationOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/DeleteAddress": {
      "delete": {
        "tags": ["Address"],
        "parameters": [
          {
            "name": "addressId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAddressOperationOutputDtoOperationOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAddressOperationOutputDtoOperationOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAddressOperationOutputDtoOperationOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/SetDefaultAddress": {
      "patch": {
        "tags": ["Address"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDefaultAddressOperationInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDefaultAddressOperationInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetDefaultAddressOperationInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SetDefaultAddressOperationOutputDtoOperationOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetDefaultAddressOperationOutputDtoOperationOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetDefaultAddressOperationOutputDtoOperationOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/UpdateAddress": {
      "patch": {
        "tags": ["Address"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressOperationInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressOperationInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressOperationInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAddressOperationOutputDtoOperationOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAddressOperationOutputDtoOperationOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAddressOperationOutputDtoOperationOutput"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/AddToCart": {
      "post": {
        "tags": ["Cart"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartOperationInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartOperationInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartOperationInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddToCartOperationOutputDtoOperationOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddToCartOperationOutputDtoOperationOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddToCartOperationOutputDtoOperationOutput"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/RemoveFromCart": {
      "delete": {
        "tags": ["Cart"],
        "parameters": [
          {
            "name": "productIds",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveFromCartOperationOutputDtoOperationOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveFromCartOperationOutputDtoOperationOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveFromCartOperationOutputDtoOperationOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/UpdateCartProduct": {
      "patch": {
        "tags": ["Cart"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartProductOperationInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartProductOperationInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartProductOperationInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCartProductOperationOutputDtoOperationOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCartProductOperationOutputDtoOperationOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCartProductOperationOutputDtoOperationOutput"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/GetPaymentMethodDetails": {
      "get": {
        "tags": ["PaymentMethods"],
        "parameters": [
          {
            "name": "methodId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaymentMethodDetailsOperationOutputDtoOperationOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaymentMethodDetailsOperationOutputDtoOperationOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaymentMethodDetailsOperationOutputDtoOperationOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/AddPaymentMethod": {
      "post": {
        "tags": ["PaymentMethods"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPaymentMethodOperationInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPaymentMethodOperationInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddPaymentMethodOperationInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddPaymentMethodOperationOutputDtoOperationOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddPaymentMethodOperationOutputDtoOperationOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddPaymentMethodOperationOutputDtoOperationOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/DeletePaymentMethod": {
      "delete": {
        "tags": ["PaymentMethods"],
        "parameters": [
          {
            "name": "methodId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeletePaymentMethodOperationOutputDtoOperationOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletePaymentMethodOperationOutputDtoOperationOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletePaymentMethodOperationOutputDtoOperationOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/SetDefaultPaymentMethod": {
      "patch": {
        "tags": ["PaymentMethods"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDefaultPaymentMethodOperationInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDefaultPaymentMethodOperationInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetDefaultPaymentMethodOperationInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SetDefaultPaymentMethodOperationOutputDtoOperationOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetDefaultPaymentMethodOperationOutputDtoOperationOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetDefaultPaymentMethodOperationOutputDtoOperationOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/UpdatePaymentMethod": {
      "patch": {
        "tags": ["PaymentMethods"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentMethodOperationInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentMethodOperationInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentMethodOperationInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePaymentMethodOperationOutputDtoOperationOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePaymentMethodOperationOutputDtoOperationOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePaymentMethodOperationOutputDtoOperationOutput"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddAddressOperationInputDto": {
        "required": [
          "city",
          "country",
          "location",
          "name",
          "phone",
          "postalCode",
          "street"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddAddressOperationOutputDto": {
        "type": "object",
        "properties": {
          "updatedAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddAddressOperationOutputDtoOperationOutput": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AddAddressOperationOutputDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OutputMetadataDto"
          }
        },
        "additionalProperties": false
      },
      "AddPaymentMethodOperationInputDto": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/PaymentMethodType"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "cardNumber": {
            "type": "string",
            "nullable": true
          },
          "expirationMonth": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "expirationYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddPaymentMethodOperationOutputDto": {
        "type": "object",
        "properties": {
          "updatedMethods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentMethodDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddPaymentMethodOperationOutputDtoOperationOutput": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AddPaymentMethodOperationOutputDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OutputMetadataDto"
          }
        },
        "additionalProperties": false
      },
      "AddToCartOperationInputDto": {
        "required": ["productIds"],
        "type": "object",
        "properties": {
          "productIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddToCartOperationOutputDto": {
        "type": "object",
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartProductDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddToCartOperationOutputDtoOperationOutput": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AddToCartOperationOutputDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OutputMetadataDto"
          }
        },
        "additionalProperties": false
      },
...
      "CartProductDto": {
        "required": ["productId"],
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isSelected": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DeleteAddressOperationOutputDto": {
        "type": "object",
        "properties": {
          "updatedAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DeleteAddressOperationOutputDtoOperationOutput": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeleteAddressOperationOutputDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OutputMetadataDto"
          }
        },
        "additionalProperties": false
      },
      "DeletePaymentMethodOperationOutputDto": {
        "type": "object",
        "properties": {
          "updatedMethods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentMethodDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DeletePaymentMethodOperationOutputDtoOperationOutput": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeletePaymentMethodOperationOutputDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OutputMetadataDto"
          }
        },
        "additionalProperties": false
      },
...
      "GetPaymentMethodDetailsOperationOutputDto": {
        "type": "object",
        "properties": {
          "method": {
            "$ref": "#/components/schemas/PaymentMethodDetailsDto"
          }
        },
        "additionalProperties": false
      },
      "GetPaymentMethodDetailsOperationOutputDtoOperationOutput": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GetPaymentMethodDetailsOperationOutputDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OutputMetadataDto"
          }
        },
        "additionalProperties": false
      },
...
      "OutputMetadataDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PaymentMethodType": {
        "enum": ["None", "Card", "Bank"],
        "type": "string"
      },
...
      "RemoveFromCartOperationOutputDto": {
        "type": "object",
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartProductDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RemoveFromCartOperationOutputDtoOperationOutput": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RemoveFromCartOperationOutputDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OutputMetadataDto"
          }
        },
        "additionalProperties": false
      },
...
      "SetDefaultAddressOperationInputDto": {
        "required": ["addressId"],
        "type": "object",
        "properties": {
          "addressId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SetDefaultAddressOperationOutputDto": {
        "type": "object",
        "properties": {
          "updatedAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SetDefaultAddressOperationOutputDtoOperationOutput": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SetDefaultAddressOperationOutputDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OutputMetadataDto"
          }
        },
        "additionalProperties": false
      },
      "SetDefaultPaymentMethodOperationInputDto": {
        "required": ["methodId"],
        "type": "object",
        "properties": {
          "methodId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SetDefaultPaymentMethodOperationOutputDto": {
        "type": "object",
        "properties": {
          "updatedMethods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentMethodDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SetDefaultPaymentMethodOperationOutputDtoOperationOutput": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SetDefaultPaymentMethodOperationOutputDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OutputMetadataDto"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateAddressOperationInputDto": {
        "required": ["id"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateAddressOperationOutputDto": {
        "type": "object",
        "properties": {
          "updatedAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateAddressOperationOutputDtoOperationOutput": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UpdateAddressOperationOutputDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OutputMetadataDto"
          }
        },
        "additionalProperties": false
      },
      "UpdateCartProductOperationInputDto": {
        "required": ["products"],
        "type": "object",
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartProductDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateCartProductOperationOutputDto": {
        "type": "object",
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartProductDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateCartProductOperationOutputDtoOperationOutput": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UpdateCartProductOperationOutputDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OutputMetadataDto"
          }
        },
        "additionalProperties": false
      },
...
      "UpdatePaymentMethodOperationInputDto": {
        "required": ["id"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "cardNumber": {
            "type": "string",
            "nullable": true
          },
          "expirationMonth": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "expirationYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdatePaymentMethodOperationOutputDtoOperationOutput": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UpdatePaymentMethodOperationOutputDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OutputMetadataDto"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}